# improve child lookup performance