Baseline performance from https://github.com/akkadotnet/akka.net/pull/5239

Done, expanded the benchmark to include some more scenarios.

``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19041.1165 (2004/May2020Update/20H1)
AMD Ryzen 7 1700, 1 CPU, 16 logical and 8 physical cores
.NET SDK=5.0.302
  [Host]     : .NET Core 3.1.17 (CoreCLR 4.700.21.31506, CoreFX 4.700.21.31502), X64 RyuJIT
  DefaultJob : .NET Core 3.1.17 (CoreCLR 4.700.21.31506, CoreFX 4.700.21.31502), X64 RyuJIT


```
|                               Method |      Mean |    Error |   StdDev |  Gen 0 | Allocated |
|------------------------------------- |----------:|---------:|---------:|-------:|----------:|
|                         ResolveChild |  64.69 ns | 0.145 ns | 0.129 ns |      - |         - |
| Resolve3DeepChildRepointableActorRef | 652.15 ns | 9.237 ns | 8.641 ns | 0.0839 |     352 B |
|       Resolve3DeepChildLocalActorRef | 311.74 ns | 4.924 ns | 4.605 ns | 0.0629 |     264 B |