@Aaronontheweb 2021-09-01
Performance from this PR:

``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19041.1165 (2004/May2020Update/20H1)
AMD Ryzen 7 1700, 1 CPU, 16 logical and 8 physical cores
.NET SDK=5.0.302
  [Host]     : .NET Core 3.1.17 (CoreCLR 4.700.21.31506, CoreFX 4.700.21.31502), X64 RyuJIT
  DefaultJob : .NET Core 3.1.17 (CoreCLR 4.700.21.31506, CoreFX 4.700.21.31502), X64 RyuJIT


```
|                               Method |      Mean |    Error |   StdDev |  Gen 0 | Allocated |
|------------------------------------- |----------:|---------:|---------:|-------:|----------:|
|                         ResolveChild |  49.62 ns | 0.034 ns | 0.029 ns |      - |         - |
| Resolve3DeepChildRepointableActorRef | 639.32 ns | 1.500 ns | 1.330 ns | 0.0839 |     352 B |
|       Resolve3DeepChildLocalActorRef | 325.46 ns | 0.653 ns | 0.545 ns | 0.0629 |     264 B |


Now accurately measuring the `Context.Child` single lookup time, eliminated `string.IndexOf` method from that call.